move comments to new component
